move common functions to conftest file
