+++ ngraph/python/tests/conftest.py
@@ -9,6 +9,31 @@ import tests
 from pathlib import Path
 
 
+def image_path():
+    path_to_repo = os.environ["DATA_PATH"]
+    path_to_img = os.path.join(path_to_repo, 'validation_set', '224x224', 'dog.bmp')
+    return path_to_img
+
+
+def model_path(is_myriad=False):
+    path_to_repo = os.environ["MODELS_PATH"]
+    if not is_myriad:
+        test_xml = os.path.join(path_to_repo, "models", "test_model", 'test_model_fp32.xml')
+        test_bin = os.path.join(path_to_repo, "models", "test_model", 'test_model_fp32.bin')
+    else:
+        test_xml = os.path.join(path_to_repo, "models", "test_model", 'test_model_fp16.xml')
+        test_bin = os.path.join(path_to_repo, "models", "test_model", 'test_model_fp16.bin')
+    return (test_xml, test_bin)
+
+
+def plugins_path():
+    path_to_repo = os.environ["DATA_PATH"]
+    plugins_xml = os.path.join(path_to_repo, 'ie_class', 'plugins.xml')
+    plugins_win_xml = os.path.join(path_to_repo, 'ie_class', 'plugins_win.xml')
+    plugins_osx_xml = os.path.join(path_to_repo, 'ie_class', 'plugins_apple.xml')
+    return (plugins_xml, plugins_win_xml, plugins_osx_xml)
+    
+
 def _get_default_model_zoo_dir():
     return Path(os.getenv("ONNX_HOME", Path.home() / ".onnx/model_zoo"))
