+++ ngraph/python/tests/test_inference_engine/test_exec_net.py
@@ -7,23 +7,7 @@ import pytest
 import warnings
 
 from openvino.inference_engine import IECore, IENetwork, ExecutableNetwork, DataPtr, InputInfoPtr, InputInfoCPtr
-
-
-def image_path():
-    path_to_repo = os.environ["DATA_PATH"]
-    path_to_img = os.path.join(path_to_repo, 'validation_set', '224x224', 'dog.bmp')
-    return path_to_img
-
-
-def model_path(is_myriad=False):
-    path_to_repo = os.environ["MODELS_PATH"]
-    if not is_myriad:
-        test_xml = os.path.join(path_to_repo, "models", "test_model", 'test_model_fp32.xml')
-        test_bin = os.path.join(path_to_repo, "models", "test_model", 'test_model_fp32.bin')
-    else:
-        test_xml = os.path.join(path_to_repo, "models", "test_model", 'test_model_fp16.xml')
-        test_bin = os.path.join(path_to_repo, "models", "test_model", 'test_model_fp16.bin')
-    return (test_xml, test_bin)
+from ..conftest import model_path, image_path
 
 
 test_net_xml, test_net_bin = model_path()
