+++ ngraph/python/tests/test_inference_engine/test_input_info.py
@@ -2,19 +2,11 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from openvino.inference_engine import IECore, IENetwork, ExecutableNetwork, InputInfoPtr, DataPtr, TensorDesc
+from ..conftest import model_path
 
 import os
 import pytest
 
-def model_path(is_myriad=False):
-    path_to_repo = os.environ["MODELS_PATH"]
-    if not is_myriad:
-        test_xml = os.path.join(path_to_repo, "models", "test_model", 'test_model_fp32.xml')
-        test_bin = os.path.join(path_to_repo, "models", "test_model", 'test_model_fp32.bin')
-    else:
-        test_xml = os.path.join(path_to_repo, "models", "test_model", 'test_model_fp16.xml')
-        test_bin = os.path.join(path_to_repo, "models", "test_model", 'test_model_fp16.bin')
-    return (test_xml, test_bin)
 
 test_net_xml, test_net_bin = model_path()
