+++ ngraph/python/tests/test_inference_engine/test_preprocessinfo.py
@@ -6,17 +6,7 @@ import os
 
 from openvino.inference_engine import PreProcessInfo, IECore, TensorDesc, Blob, \
     MeanVariant, ResizeAlgorithm, ColorFormat, PreProcessChannel
-
-
-def model_path(is_myriad=False):
-    path_to_repo = os.environ["MODELS_PATH"]
-    if not is_myriad:
-        test_xml = os.path.join(path_to_repo, "models", "test_model", 'test_model_fp32.xml')
-        test_bin = os.path.join(path_to_repo, "models", "test_model", 'test_model_fp32.bin')
-    else:
-        test_xml = os.path.join(path_to_repo, "models", "test_model", 'test_model_fp16.xml')
-        test_bin = os.path.join(path_to_repo, "models", "test_model", 'test_model_fp16.bin')
-    return (test_xml, test_bin)
+from ..conftest import model_path
 
 test_net_xml, test_net_bin = model_path()
